requirements.txt
numpy
tensorflow-gpu
stanford-corenlp